+++ src/main/java/pyre/coloredredstone/blocks/BlockColoredRedstoneLamp.java
@@ -216,6 +216,16 @@ public class BlockColoredRedstoneLamp extends Block implements IColoredFeatures,
         return 0;
     }
 
+    @Override
+    public void onFallenUpon(World worldIn, BlockPos pos, Entity entityIn, float fallDistance) {
+        if (CurrentModConfig.soft && getColor(worldIn, pos) == SOFT_COLOR && !entityIn.isSneaking()) {
+            float damageMultiplier = 1.0F - (CurrentModConfig.softDamageReduction / 100.0F);
+            entityIn.fall(fallDistance, damageMultiplier);
+        } else {
+            super.onFallenUpon(worldIn, pos, entityIn, fallDistance);
+        }
+    }
+
     @SideOnly(Side.CLIENT)
     public static int colorMultiplier(int power, EnumColor color) {
         int red = color.getShades()[power].getR();